requirements.txt
absl-py==2.3.0
albucore==0.0.24
albumentations==2.0.8
amqp==5.3.1
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
backports-datetime-fromisoformat==2.0.3
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==45.0.4
cycler==0.12.1
Cython==3.1.2
Django==4.2.23
django-cors-headers==4.7.0
django-redis==5.4.0
django-sslserver==0.22
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
easydict==1.13
environs==9.5.0
eval_type_backport==0.2.2
flatbuffers==25.2.10
flower==2.0.1
fonttools==4.58.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.14.0
humanfriendly==10.0
humanize==4.12.3
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
insightface==0.7.3
joblib==1.5.1
jsonfield==3.1.0
keras==3.10.0
kiwisolver==1.4.7
kombu==5.5.4
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
mdurl==0.1.2
milvus-lite==2.4.12
minio==7.2.15
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==0.1.1
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.18.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==5.29.5
psycopg2==2.9.10
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
pymilvus==2.3.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rich==14.0.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
simsimd==6.4.9
six==1.17.0
sqlparse==0.5.3
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tornado==6.5.1
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
